test: passing all the pest tests
