+++ src/App/Sin/Nomadelfia/GruppoFamiliare/Controllers/GruppifamiliariController.php
@@ -46,5 +46,4 @@ class GruppifamiliariController extends CoreBaseController
     }
 
 
-
 }
