+++ tests/Popolazione/Http/SinIndexTest.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace Tests\Http\Nomadelfia;
+
+use Tests\TestCase;
+
+
+it("unauthenticated_user_dont_see_all_systems", function () {
+    // laravel will not wrap the error in 500 http response, but return the raw error
+    $this->withExceptionHandling();
+
+    $this
+        ->get('/home')
+        ->assertSuccessful()
+        ->assertSee("Autenticati")
+        ->assertSee("Entra come ospite")
+        ->assertSee("Biblioteca")   //  unauthenticated users can only search books
+        ->assertDontSee("Gestione Nomadelfia")
+        ->assertDontSee("Officina")
+        ->assertDontSee("Amministratore")
+        ->assertDontSee("Agraria")
+        ->assertDontSee("Gestione Scuola")
+        ->assertDontSee("Patenti");
+});
